+++ badge/2048/render.h
@@ -8,7 +8,7 @@
 #define FIELD_SIZE 13
 
 void show_frame(badge_framebuffer * fb);
-void show_grid(badge_framebuffer * fb, uint16_t grid[GRID_SIZE][GRID_SIZE]);
+void show_grid(badge_framebuffer * fb, uint8_t grid[GRID_SIZE][GRID_SIZE]);
 void render_number(badge_framebuffer * fb, uint8_t number, uint8_t x, uint8_t y);
 void clear_field(badge_framebuffer * fb, uint8_t x_grid, uint8_t y_grid);
